加入ending
“全剧终”竟然是在等待空格，坑爹啊

+++ src/js/pal/sprite.js
@@ -41,7 +41,7 @@ utils.extend(Sprite.prototype, {
       return false;
     }
     frameNum <<= 1;
-    var offset = (this.reader.getUint16(frameNum) << 1);
+    var offset = WORD(this.reader.getUint16(frameNum) << 1);
     return offset;
   },
   /**
